# Plugin/Swagger/bundled.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API OF CUDIDI",
    "description": "This is the documentation API of Cudidi browser",
    "contact": {
      "email": "Cudidi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get me",
        "description": "Get information my user.",
        "operationId": "getMyUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "400": {
            "description": "Invalid username supplied",
            "content": {}
          },
          "401": {
            "description": "Access token is missing or invalid",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user id",
        "operationId": "getUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ID": {
                      "type": "integer",
                      "readOnly": true
                    },
                    "Username": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 30
                    },
                    "Password": {
                      "type": "string",
                      "minLength": 6,
                      "writeOnly": true
                    },
                    "FullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "Email": {
                      "type": "string",
                      "format": "email"
                    },
                    "BirthDay": {
                      "type": "string",
                      "format": "date"
                    },
                    "Address": {
                      "type": "string"
                    },
                    "Avatar": {
                      "type": "string"
                    },
                    "PhoneNumber": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 15
                    }
                  },
                  "xml": {
                    "name": "User"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ID": {
                      "type": "integer",
                      "readOnly": true
                    },
                    "Username": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 30
                    },
                    "Password": {
                      "type": "string",
                      "minLength": 6,
                      "writeOnly": true
                    },
                    "FullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "Email": {
                      "type": "string",
                      "format": "email"
                    },
                    "BirthDay": {
                      "type": "string",
                      "format": "date"
                    },
                    "Address": {
                      "type": "string"
                    },
                    "Avatar": {
                      "type": "string"
                    },
                    "PhoneNumber": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 15
                    }
                  },
                  "xml": {
                    "name": "User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updated user",
        "description": "Update information of user by id.",
        "operationId": "updateUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "the id that need to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated user object",
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "integer",
                    "readOnly": true
                  },
                  "Username": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                  },
                  "Password": {
                    "type": "string",
                    "minLength": 6,
                    "writeOnly": true
                  },
                  "FullName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "Email": {
                    "type": "string",
                    "format": "email"
                  },
                  "BirthDay": {
                    "type": "string",
                    "format": "date"
                  },
                  "Address": {
                    "type": "string"
                  },
                  "Avatar": {
                    "type": "string"
                  },
                  "PhoneNumber": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 15
                  }
                },
                "xml": {
                  "name": "User"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid user supplied",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user",
        "description": "Delete specify user by id.",
        "operationId": "deleleUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid username supplied",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
    "/user/update-password": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update password",
        "description": "Update the new password.",
        "operationId": "updatePassword",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Check password of user",
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/user/upload-avatar": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Upload avatar of user",
        "description": "Upload avatar of user by id.",
        "operationId": "uploadAvatar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Upload user object",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Invalid user supplied",
            "content": {}
          },
          "401": {
            "description": "Access token is missing or invalid",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
    "/user/update-information": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updated user",
        "description": "Update information of your user.",
        "operationId": "updateUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated user object",
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "integer",
                    "readOnly": true
                  },
                  "Username": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                  },
                  "Password": {
                    "type": "string",
                    "minLength": 6,
                    "writeOnly": true
                  },
                  "FullName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "Email": {
                    "type": "string",
                    "format": "email"
                  },
                  "BirthDay": {
                    "type": "string",
                    "format": "date"
                  },
                  "Address": {
                    "type": "string"
                  },
                  "Avatar": {
                    "type": "string"
                  },
                  "PhoneNumber": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 15
                  }
                },
                "xml": {
                  "name": "User"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid user supplied",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/article/upload-image": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Upload image of user",
        "description": "Upload image.",
        "operationId": "uploadImage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Upload image",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Invalid user supplied",
            "content": {}
          },
          "401": {
            "description": "Access token is missing or invalid",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
...
  }
}